+++ pxfish/field_type.py
@@ -1,11 +1,14 @@
 """Functions for Creating Field Types and Allowable Field Types"""
 
-
 import logging
-from definition import field_type_list
+import sample_type
+import object_type
+from definition import (
+        field_type_list
+        )
 
 
-def add_field_type(*, operation_type, definition, role, session):
+def add_field_type(*, operation_type, definition, role, path, session):
     """
     Creates list of Field Types to Add to Operation Type
     If a Field Type doesn't already exist, creates it
@@ -13,6 +16,7 @@ def add_field_type(*, operation_type, definition, role, session):
         operation_type (Operation Type): parent object
         definition (Dictionary): data about field types
         role (String): input or output
+        path (String): path to operation type
         session (Session Object): Aquarium session object
     """
     query = {
@@ -20,73 +24,106 @@ def add_field_type(*, operation_type, definition, role, session):
         'parent_id': operation_type.id,
         'role': role
     }
-    field_type = session.FieldType.where(query)
+    retrieved_field_type = session.FieldType.where(query)
 
-    if field_type:
-        ft = field_type[0]
+    if retrieved_field_type:
+        field_type = retrieved_field_type[0]
+        all_afts = field_type.allowable_field_types
+        extant_afts = field_type_list(field_types=[field_type])[0]['allowable_field_types']
+
+        for sample_obj_pair in definition['allowable_field_types']:
+            if sample_obj_pair not in extant_afts:
+                new_aft = add_aft(aft_def=sample_obj_pair, session=session, path=path)
+                all_afts.append(new_aft)
+        field_type.allowable_field_types = all_afts
     else:
-        ft = session.FieldType.new()
-        ft.role = role
-        ft.array = definition['array']
-        ft.part = definition['part']
-        ft.routing = definition['routing']
-        ft.name = query['name']
-        ft.ftype = 'sample'
-        ft.allowable_field_types = add_aft(
-            session=session, definition=definition
-        )
+        field_type = session.FieldType.new()
+        field_type.role = role
+        field_type.array = definition['array']
+        field_type.part = definition['part']
+        field_type.routing = definition['routing']
+        field_type.name = query['name']
+        field_type.ftype = definition['ftype']
 
-    return ft
+        field_type.allowable_field_types = [
+            add_aft(
+                aft_def=aft, session=session, path=path
+                )
+            for aft in definition['allowable_field_types']]
+
+    return field_type
 
 
-def build(*, operation_type, definitions, session):
+def build(*, operation_type, definitions, path, session):
     """
     Adds defined field types to Operation Type
 
     Arguments:
         operation_type (Operation Type): parent object
         definitions (Dictionary): data about field types
+        path (String): path to Operation Type
         session (Session Object): Aquarium session object
     """
+
     field_types = []
 
     for field_type_definition in definitions['inputs']:
-        field_types.append(add_field_type(
-            operation_type=operation_type,
-            definition=field_type_definition,
-            role='input', session=session)
+        field_types.append(
+            add_field_type(
+                operation_type=operation_type,
+                definition=field_type_definition,
+                role='input',
+                path=path,
+                session=session
+                )
         )
 
     for field_type_definition in definitions['outputs']:
-        field_types.append(add_field_type(
-            operation_type=operation_type,
-            definition=field_type_definition,
-            role='output', session=session)
+        field_types.append(
+            add_field_type(
+                operation_type=operation_type,
+                definition=field_type_definition,
+                role='output',
+                path=path,
+                session=session)
         )
-
     operation_type.field_types = field_types
 
     session.utils.update_operation_type(operation_type)
 
 
-def add_aft(*, session, definition):
+def add_aft(*, aft_def, session, path):
     """
     Adds Sample and Object type names to Field Type Object
 
     Arguments:
         session (Session Object): Aquarium session object
-        definition (Dictionary): Data about Sample and Object types
+        aft_def (Dictionary): Data about Sample and Object types
     """
-    afts = []
-    for aft in definition['allowable_field_types']:
-
-        sample_type = session.SampleType.new()
-        object_type = session.ObjectType.new()
-
-        sample_type.name = aft['sample_type']
-        object_type.name = aft['object_type']
-        afts.append({'sample_type': sample_type, 'object_type': object_type})
-    return afts
+    if sample_type.exists(
+            session=session,
+            sample_type=aft_def['sample_type']
+        ):
+        sampl_type = session.SampleType.new()
+        sampl_type.name = aft_def['sample_type']
+    else:
+        sampl_type = sample_type.create(
+            session=session,
+            sample_type=aft_def['sample_type'],
+            path=path
+            )
+    if object_type.exists(
+            session=session,
+            object_type=aft_def['object_type']):
+        obj_type = session.ObjectType.new()
+        obj_type.name = aft_def['object_type']
+    else:
+        obj_type = object_type.create(
+            session=session,
+            object_type=aft_def['object_type'],
+            path=path
+            )
+    return {'sample_type': sampl_type, 'object_type': obj_type}
 
 
 def equivalent(*, field_type, definition):
@@ -97,26 +134,35 @@ def equivalent(*, field_type, definition):
         definition (Dictionary): Type Data saved locally
     Returns True if equivalent, False otherwise.
     """
-    field_type_data = field_type_list(
-            field_types=[field_type], role=field_type.role)
-    for k, v in definition.items():
-        if field_type_data[0][k] != v:
+    extant_field_types = field_type_list(
+        field_types=[field_type])[0]
+
+    for key, ft_value in definition.items():
+        if key == 'allowable_field_types':
+            if len(extant_field_types['allowable_field_types']) > len(ft_value):
+                return False
+            for sample_obj_pair in extant_field_types['allowable_field_types']:
+                if sample_obj_pair not in ft_value:
+                    return False
+            continue
+
+        if extant_field_types[key] != ft_value:
             return False
     return True
 
 
-def check_for_conflicts(*, field_types, definitions):
+def check_for_conflicts(*, field_types, definitions, force):
     """
     Compares data in defintion file to data stored in an Aquarium Instance
     Arguments:
         field_types (List): Field Types retrieved from Aquarium
         defintions (Dictionary): data about field types stored in defintion file
-    returns Dictionary containing:
-        conflicts (Dictionary): Field Types in both places with differing data
-        local_diff (Dictionary): Field Types in definition file Only
-        aquarium_diff_names (Dictionary): Field Types in Aquarium Only
+        force (Boolean): if set, override conflict checks
+    returns:
+        conflicts (Dictionary): Field Types in both places with differing data (Conflict)
+        local_diff (Dictionary): Field Types in definition file Only (Ones to Push)
+        aquarium_diff_names (Dictionary): Field Types in Aquarium Only (Don't push)
     """
-
     type_map = {field_type.name: field_type for field_type in field_types}
     local_diff = dict()
     match_names = set()
@@ -126,7 +172,10 @@ def check_for_conflicts(*, field_types, definitions):
         name = definition['name']
         if name in type_map:
             match_names.add(name)
-            if not equivalent(field_type=type_map[name], definition=definition):
+            if not force and not equivalent(
+                    field_type=type_map[name],
+                    definition=definition
+            ):
                 conflicts[name] = definition
         else:
             local_diff[name] = definition
@@ -136,46 +185,60 @@ def check_for_conflicts(*, field_types, definitions):
     return (aquarium_diff_names, conflicts, local_diff)
 
 
-def types_valid(*, operation_type, definitions, session):
+def types_valid(*, operation_type, definitions, force, session):
     """
     Compares an Operation Type's Field Types to those in the Definitions file
     If a Field Type only exists in Instance Type, stops push for Operation Type
     Arguments:
         operation_type (Operation Type): parent object
         definitions (Dictionary): data about field types
+        force (Boolean): if set, overrides conflict checks
         session (Session Object): Aquarium session object
     """
-
     field_types = session.FieldType.where({'parent_id': operation_type.id})
 
     missing_inputs, input_conflicts, valid_inputs = check_for_conflicts(
         field_types=[t for t in field_types if t.role == 'input'],
-        definitions=definitions['inputs']
+        definitions=definitions['inputs'],
+        force=force
     )
     missing_outputs, output_conflicts, valid_outputs = check_for_conflicts(
         field_types=[t for t in field_types if t.role == 'output'],
-        definitions=definitions['outputs']
+        definitions=definitions['outputs'],
+        force=force
     )
 
     messages = []
+
     if missing_inputs or missing_outputs:
         for conflict in missing_inputs:
-            messages.append(f'Aquarium Field Type Input {conflict} is not in your definition file.')
+            messages.append(
+                (f'Aquarium Field Type Input, "{conflict}", '
+                 'is not in your definition file.\n')
+            )
         for conflict in missing_outputs:
-            messages.append(f'Aquarium Field Type Output {conflict} is not in your definition file.')
+            messages.append((f'Aquarium Field Type Output, "{conflict}",'
+                             'is not in your definition file.\n'))
 
     if input_conflicts or output_conflicts:
         for conflict in input_conflicts:
-            messages.append(f'There is a data conflict between the Aquarium Field Type definition of Input {conflict} and your local definition')
+            messages.append(
+                ('There is a data conflict between the Aquarium Field Type '
+                 f'Definition of Input, "{conflict}", and your local definition.\n'))
         for conflict in output_conflicts:
-            messages.append(f'There is a data conflict between the Aquarium Field Type definition of Output {conflict} and your local definition')
-
+            messages.append(
+                ('There is a data conflict between the Aquarium Field Type '
+                 f'Definition of Output, "{conflict}", and your local definition.\n')
+            )
     if messages:
         messages = ' '.join(messages)
         logging.warning(
-                'The Following Field Type Conflict(s) exist: %s. Operation Type %s will not be pushed',
-                    messages, operation_type.name
-                    )
+            'The Following Field Type Conflict(s) exist: \n %s', messages)
+        logging.warning(
+            'Operation Type "%s" will not be pushed.', operation_type.name)
+        logging.info(
+            'To override this and replace instance data with data from your definition file,\n \
+            run push again with the force flag (-f, --force) set.')
         return False
 
     return True