+++ pxfish/object_type.py
@@ -4,6 +4,7 @@ Functions for pulling Object Types in Aquarium.
 
 import json
 import logging
+import pathlib
 import os
 
 from paths import (
@@ -12,6 +13,67 @@ from paths import (
     simplename
 )
 
+
+def exists(*, session, object_type):
+    """
+    Checks whether an Object Type named in definition exists in Aquarium
+    """
+    # TODO: we are not currently storing the description in the definition file
+    object_type = session.ObjectType.where({'name': object_type})
+    return bool(object_type)
+
+
+def create(*, session, object_type, path):
+    """
+    Creates a new Object Type Type in Aquarium
+    """
+    path = pathlib.PurePath(path).parts[0]
+    path = create_named_path(path, 'object_types')
+    breakpoint()
+    try:
+        data_dict = read(path=path, object_type=object_type)
+    except FileNotFoundError:
+        return
+    # TODO: try except for File not Found Error
+    obj_type = session.ObjectType.new(
+        name=data_dict['name'],
+        description=data_dict['description'],
+        min=data_dict['min'],
+        max=data_dict['max'],
+        handler=data_dict['handler'],
+        safety=data_dict['safety'],
+        clean_up=data_dict['clean up'],
+        data=data_dict['data'],
+        vendor=data_dict['vendor'],
+        unit=data_dict['unit'],
+        cost=data_dict['cost'],
+        release_method=data_dict['release method'],
+        release_description=data_dict['release description'],
+        image=data_dict['image'],
+        prefix=data_dict['prefix'],
+        rows=data_dict['rows'],
+        columns=data_dict['columns']
+        )
+    obj_type.save()
+    return obj_type
+
+
+def read(*, path, object_type):
+    """
+    Reads the json file for the indicated object type
+    """
+    object_type = simplename(object_type) + ".json"
+    file_path = os.path.join(path, object_type)
+
+    try:
+        with open(file_path) as file:
+            breakpoint()
+            data_dict = json.load(file)
+    except FileNotFoundError as error:
+        logging.warning(
+            'Error %s reading expected code file %s', error, 'definition.json')
+        raise FileNotFoundError
+    return data_dict
 
 def write_files(*, path, object_type):
     """